First refactoring.
